requirements.txt
certifi==2024.8.30
elastic-transport==8.15.1
elasticsearch==8.16.0
Faker==33.1.0
kafka-python==2.0.2
py4j==0.10.9.7
pyspark==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
spark-nlp
textblob
vaderSentiment

